requirements.txt
DateTime==5.5
python-jenkins==1.8.2
tabulate==0.9.0
